virny_flow_interface/requirements.txt
virny_flow@git+https://github.com/denysgerasymuk799/virny-flow.git@development
gradio>=5.34.0
altair>=4.2.0
pymongo>=4.0.0
pandas>=1.5.0
plotly~=6.2.0
python-dotenv>=0.19.0